Roaming Environment Added : Field
It's pretty baron at the moment, but it's a simple example scene to run and jump around in.

Example to be added to pxlNav proper.

Roaming or Static Camera switch added and triggerable from outside of pxlNav.
That means I'll be adding a "Free-Roam" button soon, so you can just drop to the ground and run around at will.